2026-03-10-chainlit/demo/basic/requirements.txt
chainlit>=2.10.0
openai>=1.30.0
python-dotenv>=1.0.0
